chore: ensure all binaries work with --bin
