+++ bin/basectl/src/cli.rs
@@ -1,0 +1,33 @@
+//! Contains the CLI arguments for the basectl binary.
+
+use clap::{Parser, Subcommand};
+
+/// Base infrastructure control CLI.
+#[derive(Debug, Parser)]
+#[command(name = "basectl")]
+#[command(about = "Base infrastructure control CLI")]
+pub struct Cli {
+    /// Chain configuration (mainnet, sepolia, devnet, or path to config file)
+    #[arg(short = 'c', long = "config", default_value = "mainnet", global = true)]
+    pub config: String,
+
+    #[command(subcommand)]
+    pub command: Option<Commands>,
+}
+
+/// Subcommands for the basectl CLI.
+#[derive(Debug, Subcommand)]
+pub enum Commands {
+    /// Chain configuration operations
+    #[command(visible_alias = "c")]
+    Config,
+    /// Flashblocks operations
+    #[command(visible_alias = "f")]
+    Flashblocks,
+    /// DA (Data Availability) backlog monitor
+    #[command(visible_alias = "d")]
+    Da,
+    /// Command center (combined view)
+    #[command(visible_alias = "cc")]
+    CommandCenter,
+}
