+++ bin/ingress-rpc/Cargo.toml
@@ -7,6 +7,10 @@ homepage.workspace = true
 repository.workspace = true
 edition.workspace = true
 
+[[bin]]
+name = "ingress-rpc"
+path = "src/main.rs"
+
 [lints]
 workspace = true
